requirements.txt
# Core dependencies for Streamlit Cloud
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0

# Database connectivity
psycopg2-binary>=2.9.10
oracledb==3.2.0
sqlalchemy>=2.0.0

# LLM and API
openai>=1.12.0

# Data visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Excel export
openpyxl>=3.1.0
xlsxwriter>=3.1.0

# Authentication and config
PyYAML>=6.0.0
streamlit-authenticator>=0.2.3

# Environment management (for local dev)
python-dotenv>=1.0.0
